+++ src/main/java/com/iaec/guichetautomatique/entities/Retrait.java
@@ -12,7 +12,7 @@ public class Retrait extends Operation{
     public Retrait() {
     }
 
-    public Retrait(Date dateOperation, double montant, Compte compte) {
+    public Retrait(Date dateOperation, int montant, Compte compte) {
         super(dateOperation, montant, compte);
     }
 }
