+++ src/main/java/com/iaec/guichetautomatique/services/OperationService.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Page;
 public interface OperationService {
 
     public Compte consulterCompte(int numeroCompte);
-    public void verser(int numeroCompte, double montant);
-    public void retirer(int numeroCompte, double montant);
+    public void verser(int numeroCompte, int montant);
+    public void retirer(int numeroCompte, int montant);
     public Page<Operation> listOperation(int numeroCompte, int page, int size);
 }
